requirements.txt
Django==1.5.4
Fabric==1.8.0
MySQL-python==1.2.4
PIL==1.1.7
South==0.8.2
boto==2.14.0
django-pure-pagination==0.2.1
django-storages==1.1.8
ecdsa==0.9
meld3==0.6.10
paramiko==1.12.0
pycrypto==2.6
requests==2.0.0
uWSGI==1.9.18.2
wsgiref==0.1.2
